Universal Remote/whatsapp-webhook/requirements.txt
functions-framework
google-cloud-pubsub